Ensure nonempty lyrics
In normal mode (not co-composing), setting the lyrics to empty can lead to errors about a variable "token" being used before it is defined. This can be circumvented by setting lyrics to a string containing a single space. In co-composing mode this somehow does not happen in my experience, but I fixed it there anyways so the two modes don't drift apart too much.